Migrate Query.lab
