+++ packages/services/api/src/modules/lab/index.ts
@@ -1,5 +1,5 @@
 import { createModule } from 'graphql-modules';
-import { resolvers } from './resolvers';
+import { resolvers } from './resolvers.generated';
 import typeDefs from './module.graphql';
 
 export const labModule = createModule({
